3 changes/edits
The icon now gets downloaded to the data folder from the ghost.png file in the branding repo when ghost starts. If you are using Ghost on a Mac you should now receive notifications from Ghost and webhooksetup command now sets the guild icon to ghosts icon and the name to Ghost Notifications.