3 new commands and 5 changes.
Added notifications, detections and snipers commands. Edited the selfbot command to display the new commands and commands missing from there. Edited the disabledetect, enabledetect, disablesniper, enablesniper commands to be dynamic so it will work with any value in the config instead of having to manually add each detection/sniper to the commands.